Rename project
Renames project from py-vex to vexipy. py-vex is too similar to the name
pyvex, which is already registered on PyPI and prevents py-vex from
being published. Renamed the project to a hopefully clever name that has
no similar names registered on PyPI.
