--- py_vex/__init__.py
@@ -1,6 +1,0 @@
-# ruff: noqa: F401
-from py_vex.component import Component, Subcomponent
-from py_vex.document import Document
-from py_vex.statement import Statement
-from py_vex.status import StatusJustification, StatusLabel
-from py_vex.vulnerability import Vulnerability
