+++ tests/test_vulnerability.py
@@ -2,7 +2,7 @@ import pytest
 from hypothesis import given
 from hypothesis_jsonschema import from_schema
 
-from py_vex import Vulnerability
+from vexipy import Vulnerability
 
 from conftest import UPSTREAM_SCHEMA_REFS
