+++ vexipy/_util.py
