+++ vexipy/vulnerability.py
@@ -2,7 +2,7 @@ from typing import Any, Iterable, Optional, Tuple
 
 from pydantic import BaseModel, ConfigDict, Field, field_validator
 
-from py_vex._iri import Iri
+from vexipy._iri import Iri
 
 
 class Vulnerability(BaseModel):
